Update IonLeakage.R
It now automatically sets the working directory to where you copy the R file and it gives a list of values to set. 
This allows you to easily change titles, subtitles, axis and legend labels etc. 
Optimized the rest of the script and it outputs a pdf and png file automatically.
Much easier to make it work on your own data!